Title: a/b 출력, Time: 6ms, Memory: 0MB, Status: Wrong Answer - CodeTree

+++ 240821/a/b 출력/a-divide-b.c
@@ -4,6 +4,6 @@ int main() {
     // 여기에 코드를 작성해주세요.
     int a,b;
     scanf("%d %d",&a,&b);
-    printf("%.21lf", (float)a/b);
+    printf("%.21Lf", (long double)a/b);
     return 0;
 }
